Use hardcoded bash for terminal instead of $SHELL
since the SHELL variable is not necessarily set, e.g. not by default in systemd units. Also in systemd units with a defined user it is not set to the users default shell but to "/bin/sh" instead. And finally our aliases and login scripts only run fully as intended in bash.

Signed-off-by: MichaIng <micha@dietpi.com>