fix(deps): fix manifest.json path for vite v5
https://github.com/vitejs/vite/commit/74fa024

The alternative would be to update vite-embed (https://github.com/ravenclaw900/vite-embed), to use "../favicon.ico" and "../index.html" as default paths for the manifest.json. Until there is a reason for us to move the manifest.json out of the outdir root, setting the path back to what it was, is however the simpler solution.

Signed-off-by: MichaIng <micha@dietpi.com>